requirements.txt
youtube-transcript-api
deep-translator
streamlit
google-generativeai